feature: Fix remaining Long to Int in ApplicantManagementScreenImpl

+++ feature/recruitment/src/main/java/in/koreatech/koin/feature/recruitment/ui/applicantmanagement/ApplicantManagementScreen.kt
@@ -87,7 +87,7 @@ private fun ApplicantManagementScreenImpl(
     applicants: ImmutableList<Applicant>,
     modifier: Modifier = Modifier,
     onChat: () -> Unit = {},
-    onApplicantDetail: (Long) -> Unit = {}
+    onApplicantDetail: (Int) -> Unit = {}
 ) {
     LazyColumn(
         modifier = modifier.fillMaxSize(),
